add showAllOrderProduct for work with service and  service work on database and start of program in Main class

+++ hw9/task1/src/ir/maktab56/repository/impl/OrderRepositoryImpl.java
@@ -34,7 +34,7 @@ public class OrderRepositoryImpl extends BaseRepositoryImpl<Order, Long> {
     }
 
 
-    public void showAllOrderedProduct(int customerId) throws SQLException {
+    public void showAllOrderedProduct(Long customerId) throws SQLException {
 
         List<OrderDetail> list = repository.getAll();
